move base-parser.ts
