+++ src/parsers/usaa-bank-account-parser.ts
@@ -1,13 +1,13 @@
-import {readPdf} from '../readPdf';
-import {flatten2dArray} from '../util/array';
-import {dateFromSlashFormat, dateWithinRange} from '../util/date';
-import {collapseSpaces, sanitizeNumberString} from '../util/string';
 import {
     createParserStateMachine,
     ParsedOutput,
     ParsedTransaction,
     StatementParser,
-} from './base-parser';
+} from '../parser-base/base-parser';
+import {readPdf} from '../readPdf';
+import {flatten2dArray} from '../util/array';
+import {dateFromSlashFormat, dateWithinRange} from '../util/date';
+import {collapseSpaces, sanitizeNumberString} from '../util/string';
 
 enum State {
     PAGE_HEADER = 'header',
